# docs/swagger.yaml
basePath: /v1
definitions:
...
  swagger.ForbiddenError:
    properties:
      error:
        example: forbidden
        type: string
    type: object
  swagger.ForbiddenResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ForbiddenError'
        type: array
    type: object
...
  swagger.GarminStatusResponse:
    properties:
      connected:
        example: true
        type: boolean
      data:
        example: true
        type: boolean
    type: object
  swagger.GarminTokenDetails:
    properties:
      access_token:
        example: a1f2b3c4-d5e6-7890-1234-56789abcdef0
        type: string
      access_token_secret:
        example: s3cr3tT0kenValu3XyZ123
        type: string
      garmin_user_id:
        example: d3a1c9e8-7b2f-4c99-b77e-8e67cd1a2b10
        type: string
    type: object
  swagger.GarminTokenExistsResponse:
    properties:
      exists:
        example: true
        type: boolean
    type: object
  swagger.GarminTokenInput:
    properties:
      details:
        $ref: '#/definitions/swagger.GarminTokenDetails'
      user_id:
        example: e19c1832-d7f3-4d65-90ea-33a3d7f6d6df
        type: string
    type: object
...
  swagger.GarminUserIDResponse:
    properties:
      user_id:
        example: e19c1832-d7f3-4d65-90ea-33a3d7f6d6df
        type: string
    type: object
...
  swagger.InternalServerError:
    properties:
      error:
        example: the server encountered a problem
        type: string
    type: object
  swagger.InternalServerErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.InternalServerError'
        type: array
    type: object
...
  swagger.ServiceUnavailableError:
    properties:
      error:
        example: database is unavailable
        type: string
    type: object
  swagger.ServiceUnavailableResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ServiceUnavailableError'
        type: array
    type: object
...
  swagger.UnauthorizedError:
    properties:
      error:
        example: unauthorized
        type: string
    type: object
  swagger.UnauthorizedResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.UnauthorizedError'
        type: array
    type: object
...
  swagger.ValidationError:
    properties:
      parameter name:
        example: full description of the error
        type: string
    type: object
  swagger.ValidationErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ValidationError'
        type: array
    type: object
info:
  contact: {}
  description: API for integrating, analyzing, and visualizing sports and health data
  termsOfService: https://csc.fi/en/security-privacy-data-policy-and-open-source-policy/privacy/
  title: KUHA REST API
paths:
...
  /utv/garmin/status:
    get:
      consumes:
      - application/json
      description: Returns whether a user has connected their Garmin account and whether
        any Garmin data exists
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.GarminStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Check Garmin connection & data status
      tags:
      - UTV - Garmin
  /utv/garmin/token:
    post:
      consumes:
      - application/json
      description: Saves or updates the Garmin token for a user
      parameters:
      - description: Garmin token input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/swagger.GarminTokenInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Save or update Garmin token
      tags:
      - UTV - Garmin
  /utv/garmin/token-exists:
    get:
      consumes:
      - application/json
      description: Checks whether a given Garmin access token is stored in the system
      parameters:
      - description: Garmin access token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.GarminTokenExistsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Check if Garmin token exists
      tags:
      - UTV - Garmin
...
  /utv/garmin/user-id-by-token:
    get:
      consumes:
      - application/json
      description: Returns the user ID associated with a given Garmin access token
      parameters:
      - description: Garmin access token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.GarminUserIDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Get user_id by Garmin access token
      tags:
      - UTV - Garmin
...
securityDefinitions:
  BearerAuth:
    description: 'Use format: Bearer your_JWT_here'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"